[B-1019] Implement missing REST endpoints (#19)

+++ application/rest.api/src/main/java/com/onlydust/marketplace/indexer/rest/api/PullRequestsRestApi.java
@@ -14,8 +14,8 @@ public class PullRequestsRestApi implements PullRequestApi {
     private final IndexingService onDemandIndexer;
 
     @Override
-    public ResponseEntity<Void> indexPullRequest(String repoOwner, String repoName, Integer pullRequestNumber) {
-        onDemandIndexer.indexPullRequest(repoOwner, repoName, pullRequestNumber.longValue());
+    public ResponseEntity<Void> indexPullRequest(String repoOwner, String repoName, Long pullRequestNumber) {
+        onDemandIndexer.indexPullRequest(repoOwner, repoName, pullRequestNumber);
         return ResponseEntity.noContent().build();
     }
 }
